12 - Reloads Modules before looking for help files
When a module doesn't live in the PSModulePath then the
module isn't loaded when the Test-Command method looks for the
help data. This change will Import the module if it's not currently
imported before running the test for the help files which allows modules
being developed outside the path to still run the tests.